requirements.txt
MySQL-python==1.2.5
Django==1.11.4
django-filter==1.0.4
